Player Container Update
Added code to generate Player Containers upon successful Client connections. These containers are assigned to the players that successfully log into the game, and are populated with dummy JSON stats. The Player Container data can then be sent to the Client for visualization.

I also put the IssuerWebsocket and the MintAndOfferCallouts into their own Scenes.